Refactored stages to only be capable of displaying 1 for per stage
We no longer require this use case for the stages class so have
removed it. Also moved the save code for the base class on PleaStage
save into the PleaStage class so it's self contained and can be nojs'd
properly.

[MAPDEV286]
