+++ apps/plea/forms.py
@@ -42,14 +42,8 @@ class NoJSPleaStepForm(BasePleaStepForm):
         try:
             self.data = args[0]
         except IndexError:
-            self.data = {}
+            self.data = kwargs.get("data", {})
 
-        # Multiple forms store their data in kwargs, it seems
-        if not self.data:
-            kwargs_data = kwargs.get("data", None)
-            if kwargs_data:
-                self.data = kwargs_data
-        
         if hasattr(self, "dependencies"):
             prefix = kwargs.get("prefix", None)
             self.check_dependencies(self.dependencies, prefix)
