+++ apps/plea/views.py
@@ -7,7 +7,7 @@ from django.views.generic import TemplateView, FormView
 
 from brake.decorators import ratelimit
 
-from apps.govuk_utils.forms import MultiStageForm
+from apps.govuk_utils.stages import MultiStageForm
 
 from .models import Case, Court
 from .forms import CourtFinderForm
